Change data variable ID integer to long

+++ src/main/java/com/limt/Models/AdminLibrarian.java
@@ -3,16 +3,16 @@ package com.limt.Models;
 import org.jetbrains.annotations.NotNull;
 
 public class AdminLibrarian extends User{
-    private Integer adminLibrarianId;
+    private Long adminLibrarianId;
 
     public AdminLibrarian(@NotNull User user){
         super(user.getUserName(), user.getPassword(), user.getEmail(), user.getPhoneNumber(), user.getAddressLine(), user.getDateOfBirth());
     }
 
-    public Integer getAdminLibrarianId() {
+    public Long getAdminLibrarianId() {
         return adminLibrarianId;
     }
-    public void setAdminLibrarianId(Integer adminLibrarianId) {
+    public void setAdminLibrarianId(Long adminLibrarianId) {
         this.adminLibrarianId = adminLibrarianId;
     }
 }
